Move lettings_site django app view files and setting urls up
